% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld.R
\name{create_polygon_fills}
\alias{create_polygon_fills}
\title{Create SLD polygon fills}
\usage{
create_polygon_fills(data, value_column, geometry_column, label_column,
  n = 5, pal = "Blues", opacity = 0.7, title = column,
  abstract = column, max_scale = 6e+05, font_family = "Arial",
  font_size = 9, font_style = "normal", font_weight = "bold",
  font_color = "#000000", halo_size = 1.5, halo_color = "#FFFFFF")
}
\description{
Create SLD polygon fills
}
