% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_build}
\alias{docker_build}
\title{Build Docker container}
\usage{
docker_build(tag = "geoserver")
}
\arguments{
\item{tag}{name to give Docker image}
}
\description{
Build container from provided Dockerfile.  No guarantee is provided this will
work without issue!  Especially on Windows...
}
