% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_run}
\alias{docker_run}
\title{Run Docker container}
\usage{
docker_run(image = "geoserver", name = "geoserver", port = 8080)
}
\arguments{
\item{image}{name of docker image}

\item{name}{name to give running instance}

\item{port}{local port}
}
\description{
Run container built via \code{docker_build}.  No guarantee is provided this
will work without issue!  Especially on Windows...
}
