% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld.R
\name{fill_rule}
\alias{fill_rule}
\title{Polygon fill rule for sld}
\usage{
fill_rule(op = c("<", "<=", ">", ">=", "==", "--"), column, value, color,
  opacity = 0.7, title)
}
\description{
Polygon fill rule for sld
}
