% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver.R
\docType{data}
\name{GeoServer}
\alias{GeoServer}
\title{GeoServer class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeoServer
}
\description{
GeoServer class
}
\keyword{datasets}
