% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld.R
\name{import_template}
\alias{import_template}
\title{Import sld template}
\usage{
import_template(template, ...)
}
\arguments{
\item{template}{name of template}

\item{...}{named arguments}
}
\description{
Import sld template
}
\details{
For every named argument, \code{name = value}, replace \code{\{\{ name \}\}}
in \code{styles/<template.sld} with \code{value}.
}
\examples{
import_template(
  "outline_with_label",
  name = "mb2018outline", strokeColor = "#000000", strokeWidth = 1,
  maxScale = 600000, geometryName = "geom", labelName = "code",
  fontFamily = "Arial", fonSize = 10, fontStyle = "normal",
  fontWeight = "bold", fontColor = "#000000"
)
}
